feat: added new feats[1]

+++ src/routes/histories-routes.ts
@@ -4,6 +4,8 @@ import HistoriesService from '../service/histories-service.js'
 import WorkersService from '../service/workers-service.js'
 import { v4 as uuidV4 } from 'uuid'
 import { historySchema } from '../model/historySchema.js'
+import checkToken from '../helpers/checkToken.js'
+import checkPermission, { permissionIndex } from '../helpers/checkPermission.js'
 
 const router = Router()
 const workersService = new WorkersService('./workers.json')
@@ -12,14 +14,26 @@ const historiesService = new HistoriesService(
   workersService
 )
 // @ts-ignore
-router.post( '/history',(req: Request<{}, {}, ICreateHistory>, res: Response) => {
+router.post(
+  '/history',
+  // @ts-ignore
+  checkToken,
+  checkPermission([
+    permissionIndex.admin,
+    permissionIndex.director,
+    permissionIndex.manager
+  ]),
+  (req: Request, res: Response) => {
     const newHistory: IHistory = req.body
-    const result = historySchema.safeParse(newHistory) 
+    const result = historySchema.safeParse(newHistory)
     if (!result.success) {
       return res.status(400).json({ errors: result.error.errors })
     }
 
-    const addedHistory = historiesService.addHistory({ ...newHistory, id: uuidV4() })
+    const addedHistory = historiesService.addHistory({
+      ...newHistory,
+      id: uuidV4()
+    })
     if (addedHistory) {
       res.status(201).json({ message: 'History added successfully' })
     } else {
@@ -28,25 +42,55 @@ router.post( '/history',(req: Request<{}, {}, ICreateHistory>, res: Response) =>
   }
 )
 
-router.delete('/history/:id', (req, res) => {
-  const isDeleted = historiesService.deleteHistory(req.params.id)
-  if (isDeleted) {
-    res.status(201).json({ message: 'History added successfully' })
-  } else {
-    res.status(404).json({ message: 'History not found' })
+router.delete(
+  '/history/:id',
+  // @ts-ignore
+  checkToken,
+  checkPermission([
+    permissionIndex.admin,
+    permissionIndex.director,
+    permissionIndex.manager
+  ]),
+  (req, res) => {
+    const isDeleted = historiesService.deleteHistory(req.params.id)
+    if (isDeleted) {
+      res.status(201).json({ message: 'History added successfully' })
+    } else {
+      res.status(404).json({ message: 'History not found' })
+    }
   }
-})
+)
 
-router.get('/histories', (req: Request, res: Response) => {
-  try {
-    const histories = historiesService.getHistories()
-    res.json(histories)
-  } catch (error) {
-    res.status(500).json({ error: 'Не удалось получить истории' })
+router.get(
+  '/histories',
+  // @ts-ignore
+  checkToken,
+  checkPermission([
+    permissionIndex.admin,
+    permissionIndex.director,
+    permissionIndex.manager,
+    permissionIndex.user
+  ]),
+  (req: Request, res: Response) => {
+    try {
+      const histories = historiesService.getHistories()
+      res.json(histories)
+    } catch (error) {
+      res.status(500).json({ error: 'Не удалось получить истории' })
+    }
   }
-})
-// @ts-ignore
-router.get('/histories/worker/:workerId',(req: Request<{ workerId: string }>, res: Response) => {
+)
+router.get(
+  '/histories/worker/:workerId',
+  // @ts-ignore
+  checkToken,
+  checkPermission([
+    permissionIndex.admin,
+    permissionIndex.director,
+    permissionIndex.manager
+  ]),
+  // @ts-ignore
+  (req: Request<{ workerId: string }>, res: Response) => {
     const workerId = req.params.workerId
 
     if (!workerId) {
